Back-app/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-environ==0.11.2
django-cors-headers==4.6.0
django-filter==24.3
channels==4.2.0
channels-redis==4.2.1
celery==5.4.0
redis==5.2.1
Pillow==11.1.0
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.25.3
coverage==7.6.10
factory_boy==3.3.1
Faker==33.1.0
bandit==1.9.4
daphne==4.1.2
